requirements.txt
pandas==1.0.4
Pillow==7.0.0
tensorflow==2.2.0
numpy==1.18.5
matplotlib==3.0.2 
keras==2.3.1
scipy==1.4.1
